add header to disconnected state
